Add support for creating neutron networks
Add support for creating Neutron networks/subnets as part of
each tenant context, allowing instances in OpenStack
deployments based on Neutron with per-tenant based overlay
networks to be tested with an actual nic.

Right now this is relatively simplistic in that each subnet
created has the same CIDR, so Neutron must be configured
to allow overlapping IP ranges in tenant subnets.

Change-Id: Ie0b3f9b72a55e8bf783f7ebf80a3287a7daf5e66